Bug 1515052 - Use WeakPtr for extension parent pointer. r=lsalzman a=jcristau
Differential Revision: https://phabricator.services.mozilla.com/D31942

--HG--
extra : source : efb03ccd9427d558e2aabb596d26e70301151031
extra : histedit_source : f4afeed7aed9b4cc39275901bf7c284a100dc699

+++ dom/canvas/WebGLExtensionVertexArray.cpp
@@ -20,26 +20,26 @@ WebGLExtensionVertexArray::~WebGLExtensionVertexArray() {}
 
 already_AddRefed<WebGLVertexArray>
 WebGLExtensionVertexArray::CreateVertexArrayOES() {
-  if (mIsLost) return nullptr;
+  if (mIsLost || !mContext) return nullptr;
 
   return mContext->CreateVertexArray();
 }
 
 void WebGLExtensionVertexArray::DeleteVertexArrayOES(WebGLVertexArray* array) {
-  if (mIsLost) return;
+  if (mIsLost || !mContext) return;
 
   mContext->DeleteVertexArray(array);
 }
 
 bool WebGLExtensionVertexArray::IsVertexArrayOES(
     const WebGLVertexArray* array) {
-  if (mIsLost) return false;
+  if (mIsLost || !mContext) return false;
 
   return mContext->IsVertexArray(array);
 }
 
 void WebGLExtensionVertexArray::BindVertexArrayOES(WebGLVertexArray* array) {
-  if (mIsLost) return;
+  if (mIsLost || !mContext) return;
 
   mContext->BindVertexArray(array);
 }
